Make logging consistent
